[ISSUE-3] ViewModelInjectionComponent 제거, 관련 불필요 리소스 제거
