--- PRACTICE_JETPACK/app/src/main/java/nlab/practice/jetpack/di/android/InjectableAndroidViewModel.kt
@@ -1,27 +1,0 @@
-package nlab.practice.jetpack.di.android
-
-import android.app.Application
-import androidx.lifecycle.AndroidViewModel
-import nlab.practice.jetpack.util.di.AppComponent
-import nlab.practice.jetpack.di.component.ViewModelInjectComponent
-
-/**
- * 의존성 주입을 지원하는 AndroidViewModel
- *
- * @author Doohyun
- * @since 2018. 11. 23
- */
-abstract class InjectableAndroidViewModel(application: Application) : AndroidViewModel(application) {
-
-    protected val injector: ViewModelInjectComponent by lazy {
-        if (application is AppComponent.Supplier) {
-            (application as AppComponent.Supplier)
-                    .getAppComponent()
-                    .viewModelComponent()
-                    .build()
-        } else {
-            throw Throwable("Error by none support AppComponent.Supplier")
-        }
-    }
-
-}
