--- PRACTICE_JETPACK/app/src/main/java/nlab/practice/jetpack/di/scope/ViewModelLifeCycle.kt
@@ -1,12 +1,0 @@
-package nlab.practice.jetpack.di.scope
-
-import javax.inject.Scope
-
-/**
- * ViewModel 에 대한 LifeCycle 과 동일시 하기 위한 Scope
- *
- * @author Doohyun
- * @since 2018. 11. 23
- */
-@Scope
-annotation class ViewModelLifeCycle
