--- PRACTICE_JETPACK/app/src/main/java/nlab/practice/jetpack/ui/common/MainContainerViewModel.kt
@@ -1,15 +1,0 @@
-package nlab.practice.jetpack.ui.common
-
-import nlab.practice.jetpack.util.databinding.model.RecyclerViewConfig
-import nlab.practice.jetpack.util.recyclerview_de.anko.AnkoViewBindingItem
-
-/**
- * @author Doohyun
- * @since 2018. 12. 12
- */
-interface MainContainerViewModel {
-    fun getItems(): List<AnkoViewBindingItem>? = null
-    fun getHeader(): AnkoViewBindingItem? = null
-    fun getFooter(): AnkoViewBindingItem? = null
-    fun getRecyclerViewConfig(): RecyclerViewConfig? = null
-}
