+++ PRACTICE_JETPACK/app/src/main/java/nlab/practice/jetpack/ui/tutorial/ActivityAnkoFirstUI.kt
@@ -11,8 +11,8 @@ import androidx.annotation.StringRes
 import androidx.constraintlayout.widget.ConstraintLayout.LayoutParams.*
 import nlab.practice.jetpack.R
 import nlab.practice.jetpack.util.color
-import nlab.practice.jetpack.util.databinding.DataBindingAnkoComponent
-import nlab.practice.jetpack.util.databinding.binder.onClick
+import nlab.practice.jetpack.util.anko.DataBindingAnkoComponent
+import nlab.practice.jetpack.util.anko.binder.onClick
 import nlab.practice.jetpack.util.string
 import org.jetbrains.anko.*
 import org.jetbrains.anko.constraint.layout.constraintLayout
