+++ PRACTICE_JETPACK/app/src/main/java/nlab/practice/jetpack/util/anko/callback/OnPropertyChangedCallbackManager.kt
@@ -1,4 +1,4 @@
-package nlab.practice.jetpack.util.databinding.callback
+package nlab.practice.jetpack.util.anko.callback
 
 import androidx.databinding.BaseObservable
 import androidx.databinding.Observable
