+++ PRACTICE_JETPACK/app/src/main/java/nlab/practice/jetpack/util/di/activity/ActivityCommonModule.kt
@@ -6,7 +6,6 @@ import dagger.Provides
 import io.reactivex.disposables.CompositeDisposable
 import nlab.practice.jetpack.util.ActivityStarterUsecase
 import nlab.practice.jetpack.util.createLazyCompositeDisposable
-import nlab.practice.jetpack.util.di.itemview.ItemViewModelFactoryModule
 import nlab.practice.jetpack.util.lifecycle.ActivityLifeCycleBinder
 import nlab.practice.jetpack.util.lifecycle.LifeCycleBinder
 
@@ -14,9 +13,7 @@ import nlab.practice.jetpack.util.lifecycle.LifeCycleBinder
  * @author Doohyun
  * @since 2018. 12. 18
  */
-@Module(includes = [
-    ItemViewModelFactoryModule::class
-])
+@Module
 class ActivityCommonModule {
 
     @ActivityScope
