--- PRACTICE_JETPACK/app/src/main/java/nlab/practice/jetpack/util/di/itemview/ItemDecorationModule.kt
@@ -1,19 +1,0 @@
-package nlab.practice.jetpack.util.di.itemview
-
-import android.content.Context
-import dagger.Module
-import dagger.Provides
-import dagger.Reusable
-import nlab.practice.jetpack.ui.home.HomeItemDelDecoration
-
-/**
- * @author Doohyun
- * @since 2018. 12. 13
- */
-@Module
-class ItemDecorationModule {
-
-    @Reusable
-    @Provides
-    fun provideHomeDecoration(context: Context) = HomeItemDelDecoration(context)
-}
