--- PRACTICE_JETPACK/app/src/main/java/nlab/practice/jetpack/util/di/itemview/ItemViewModelFactoryModule.kt
@@ -1,11 +1,0 @@
-package nlab.practice.jetpack.util.di.itemview
-
-import dagger.Module
-
-/**
- * @author Doohyun
- */
-@Module(includes = [
-
-])
-class ItemViewModelFactoryModule
