+++ PRACTICE_JETPACK/app/src/main/java/nlab/practice/jetpack/util/di/itemview/ItemViewUsecaseFactory.kt
@@ -20,7 +20,5 @@ interface ItemViewUsecaseFactory {
         fun build(): ItemViewUsecaseFactory
     }
 
-    fun testItemViewUsecase(): TestItemViewUsecase
-
     fun navigateViewUsecase(): NavigateViewUsecase
 }
