--- PRACTICE_JETPACK/app/src/main/java/nlab/practice/jetpack/util/di/itemview/TestItemViewUsecase.kt
@@ -1,10 +1,0 @@
-package nlab.practice.jetpack.util.di.itemview
-
-import android.view.View
-import javax.inject.Inject
-
-/**
- * @author Doohyun
- */
-@ItemViewScope
-class TestItemViewUsecase @Inject constructor(view: View)
