--- PRACTICE_JETPACK/app/src/main/java/nlab/practice/jetpack/util/viewmodel/InjectableViewModelFactory.kt
@@ -1,15 +1,0 @@
-package nlab.practice.jetpack.util.viewmodel
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-
-/**
- * @author Doohyun
- * @since 2018. 12. 18
- */
-class InjectableViewModelFactory : ViewModelProvider.Factory {
-
-    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-    }
-}
